% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{cleanREPV2}
\alias{cleanREPV2}
\title{Title of cleanREPV2}
\usage{
cleanREPV2(y, gen, fam = NULL, thr = 0.95)
}
\arguments{
\item{y}{Description of y.}

\item{gen}{Description of gen.}

\item{fam=NULL}{Description of fam=NULL.}

\item{thr=0.95}{Description of thr=0.95.}
}
\value{
Description of what is returned.
}
\description{
Description of what cleanREPV2 does.
}
\examples{
\dontrun{
# Example usage of cleanREPV2
result <- cleanREPV2(...)
}
}
