% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enviromics_Functions.R
\name{extract_GIS_terra}
\alias{extract_GIS_terra}
\title{Title of extract_GIS_terra}
\usage{
extract_GIS_terra(
  covraster = NULL,
  Latitude = NULL,
  Longitude = NULL,
  env.data = NULL,
  env.id = NULL,
  name.out = NULL
)
}
\arguments{
\item{covraster}{= NULL Description of covraster = NULL.}

\item{Latitude}{= NULL Description of Latitude = NULL.}

\item{Longitude}{= NULL Description of Longitude = NULL.}

\item{env.data}{= NULL Description of env.data = NULL.}

\item{env.id}{= NULL Description of env.id = NULL.}

\item{name.out}{= NULL Description of name.out = NULL.}
}
\value{
Description of what is returned.
}
\description{
Description of what extract_GIS_terra does.
}
\examples{
\dontrun{
# Example usage of extract_GIS_terra
result <- extract_GIS_terra(...)
}
}
