% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Functions.R
\name{fitMEModels_CV_fits}
\alias{fitMEModels_CV_fits}
\title{Fit multi-environment genomic prediction models for cross validation schemes}
\usage{
fitMEModels_CV_fits(
  DT,
  genoDat,
  strainGeno,
  KG = NULL,
  KE = NULL,
  Fixed,
  CVMet,
  k,
  nIter,
  FactVar,
  method,
  Reaction = FALSE,
  dimension_KE = NULL,
  bandwidth = 1,
  quantil = 0.5,
  trait = NULL,
  store_full_fits = FALSE,
  b_iter = 1000,
  b_burnin = 200,
  b_thin = 10,
  b_tol = 1e-10,
  b_R2 = 0.5,
  b_digits = 4
)
}
\arguments{
\item{DT}{Data frame with phenotypic data.}

\item{genoDat}{data frame with genotypic data.}

\item{strainGeno}{Character vector containing strain IDs of genotypes.}

\item{KG}{Matrix or NULL. Genotypic relationship matrix. Default = NULL, in which case
the kernel is estimated internally. A user-defined matrix can be supplied.}

\item{KE}{Matrix or NULL. Environmental relationship matrix. Default = NULL.
Required if `FitEnvModels = TRUE`.}

\item{CVMet}{Character. Cross-validation scheme. Default = "CV1".
Options: "CV1", "CV2", "CV0", "CV00","CV-LOFO".}

\item{method}{Character. Kernel method for estimating genotypic and
genotype × environment kernels. Default = "Linear". Option: "Gaussian".}

\item{trait}{Character. Variable specifying the trait name.}

\item{store_full_fits}{logical value to store full fits in results. Default value is set to FALSE.}

\item{K}{Integer. Number of folds in k-fold cross-validation.}

\item{NIter}{Integer. Number of replicates for cross-validation.}

\item{fitEnvModels}{Logical. Whether to include environmental covariates in the model.
Default = FALSE. If TRUE, a user-provided `KE` must be supplied.}
}
\value{
Description of what is returned.
}
\description{
Fit multi-environment genomic prediction models for cross validation schemes
}
\examples{
\dontrun{
# Example usage of fitMEModels_CV_fits
result <- fitMEModels_CV_fits(...)
}
}
