% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Functions.R
\name{fitMEModels_LOF_CV}
\alias{fitMEModels_LOF_CV}
\title{Title of fitMEModels_LOF_CV}
\usage{
fitMEModels_LOF_CV(
  DT,
  genoDat,
  strainGeno,
  KG = NULL,
  KE = NULL,
  factVar,
  factr,
  method,
  fitEnvModels = FALSE,
  FixedTerm = Fixed,
  IDColsList = IDColsMEList
)
}
\arguments{
\item{DT}{Description of DT.}

\item{genoDat}{Description of genoDat.}

\item{strainGeno}{Description of strainGeno.}

\item{factVar}{Description of factVar.}

\item{factr}{Description of factr.}

\item{method}{Description of method.}

\item{KG=NULL}{Description of KG=NULL.}

\item{KE=NULL}{Description of KE=NULL.}

\item{fitEnvModels=FALSE}{Description of fitEnvModels=FALSE.}

\item{FixedTerm=Fixed}{Description of FixedTerm=Fixed.}

\item{IDColsList=IDColsMEList}{Description of IDColsList=IDColsMEList.}
}
\value{
Description of what is returned.
}
\description{
Description of what fitMEModels_LOF_CV does.
}
\examples{
\dontrun{
# Example usage of fitMEModels_LOF_CV
result <- fitMEModels_LOF_CV(...)
}
}
