% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Functions.R
\name{get_geno_kernel}
\alias{get_geno_kernel}
\title{Title of get_geno_kernel}
\usage{
get_geno_kernel(
  Y,
  X,
  KG,
  KE,
  intercept.random = FALSE,
  model,
  method,
  dimension_KE = NULL
)
}
\arguments{
\item{Y}{Description of Y.}

\item{X}{Description of X.}

\item{KG}{Description of KG.}

\item{KE}{Description of KE.}

\item{model}{Description of model.}

\item{method}{Description of method.}

\item{intercept.random=FALSE}{Description of intercept.random=FALSE.}

\item{dimension_KE=NULL}{Description of dimension_KE=NULL.}
}
\value{
Description of what is returned.
}
\description{
Description of what get_geno_kernel does.
}
\examples{
\dontrun{
# Example usage of get_geno_kernel
result <- get_geno_kernel(...)
}
}
