% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enviromics_Functions.R
\name{get_helper}
\alias{get_helper}
\title{Title of get_helper}
\usage{
get_helper(lon, lat, variables.names, start.day, end.day, env.id, save)
}
\arguments{
\item{lon}{Description of lon.}

\item{lat}{Description of lat.}

\item{variables.names}{Description of variables.names.}

\item{start.day}{Description of start.day.}

\item{end.day}{Description of end.day.}

\item{env.id}{Description of env.id.}

\item{save}{Description of save.}
}
\value{
Description of what is returned.
}
\description{
Description of what get_helper does.
}
\examples{
\dontrun{
# Example usage of get_helper
result <- get_helper(...)
}
}
