% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Functions.R
\name{get_kernel_MDe}
\alias{get_kernel_MDe}
\title{Title of get_kernel_MDe}
\usage{
get_kernel_MDe(Y, KG, KE, intercept.random = FALSE, dimension_KE = NULL)
}
\arguments{
\item{Y}{Description of Y.}

\item{KG}{Description of KG.}

\item{KE}{Description of KE.}

\item{intercept.random=FALSE}{Description of intercept.random=FALSE.}

\item{dimension_KE=NULL}{Description of dimension_KE=NULL.}
}
\value{
Description of what is returned.
}
\description{
Description of what get_kernel_MDe does.
}
\examples{
\dontrun{
# Example usage of get_kernel_MDe
result <- get_kernel_MDe(...)
}
}
