% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getCombinedTab}
\alias{getCombinedTab}
\title{Title of getCombinedTab}
\usage{
getCombinedTab(outputListME, TraitME, IDColsME, IDColME, fitEnvCov)
}
\arguments{
\item{outputListME}{Description of outputListME.}

\item{TraitME}{Description of TraitME.}

\item{IDColsME}{Description of IDColsME.}

\item{IDColME}{Description of IDColME.}

\item{fitEnvCov}{Description of fitEnvCov.}
}
\value{
Description of what is returned.
}
\description{
Description of what getCombinedTab does.
}
\examples{
\dontrun{
# Example usage of getCombinedTab
result <- getCombinedTab(...)
}
}
