% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enviromics_Functions.R
\name{getEnvData}
\alias{getEnvData}
\title{Title of getEnvData}
\usage{
getEnvData(Coord_Data_Tab, startDate, endDate)
}
\arguments{
\item{Coord_Data_Tab}{Description of Coord_Data_Tab.}

\item{startDate}{Description of startDate.}

\item{endDate}{Description of endDate.}
}
\value{
Description of what is returned.
}
\description{
Description of what getEnvData does.
}
\examples{
\dontrun{
# Example usage of getEnvData
result <- getEnvData(...)
}
}
