% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enviromics_Functions.R
\name{getEnvKernel}
\alias{getEnvKernel}
\title{Title of getEnvKernel}
\usage{
getEnvKernel(Env.Data, process = FALSE, Gaussian = FALSE)
}
\arguments{
\item{Env.Data}{Description of Env.Data.}

\item{process=FALSE}{Description of process=FALSE.}

\item{Gaussian=FALSE}{Description of Gaussian=FALSE.}
}
\value{
Description of what is returned.
}
\description{
Description of what getEnvKernel does.
}
\examples{
\dontrun{
# Example usage of getEnvKernel
result <- getEnvKernel(...)
}
}
