% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getFilteredSitesGenoData}
\alias{getFilteredSitesGenoData}
\title{Title of getFilteredSitesGenoData}
\usage{
getFilteredSitesGenoData(tasGeno, siteMinCnt, MAF)
}
\arguments{
\item{tasGeno}{Description of tasGeno.}

\item{siteMinCnt}{Description of siteMinCnt.}

\item{MAF}{Description of MAF.}
}
\value{
Description of what is returned.
}
\description{
Description of what getFilteredSitesGenoData does.
}
\examples{
\dontrun{
# Example usage of getFilteredSitesGenoData
result <- getFilteredSitesGenoData(...)
}
}
