% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getFilteredTaxaGenoData}
\alias{getFilteredTaxaGenoData}
\title{Title of getFilteredTaxaGenoData}
\usage{
getFilteredTaxaGenoData(tasGeno, MinNotMissing)
}
\arguments{
\item{tasGeno}{Description of tasGeno.}

\item{MinNotMissing}{Description of MinNotMissing.}
}
\value{
Description of what is returned.
}
\description{
Description of what getFilteredTaxaGenoData does.
}
\examples{
\dontrun{
# Example usage of getFilteredTaxaGenoData
result <- getFilteredTaxaGenoData(...)
}
}
