% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getFixedData_List}
\alias{getFixedData_List}
\title{Title of getFixedData_List}
\usage{
getFixedData_List(
  Processed_Data_Table_Num_Split_Filt,
  trait,
  fixedCoVar,
  target_ID
)
}
\arguments{
\item{Processed_Data_Table_Num_Split_Filt}{Description of Processed_Data_Table_Num_Split_Filt.}

\item{trait}{Description of trait.}

\item{fixedCoVar}{Description of fixedCoVar.}

\item{target_ID}{Description of target_ID.}
}
\value{
Description of what is returned.
}
\description{
Description of what getFixedData_List does.
}
\examples{
\dontrun{
# Example usage of getFixedData_List
result <- getFixedData_List(...)
}
}
