% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getFreq_Alleles}
\alias{getFreq_Alleles}
\title{Title of getFreq_Alleles}
\usage{
getFreq_Alleles(GenoTable)
}
\arguments{
\item{GenoTable}{Description of GenoTable.}
}
\value{
Description of what is returned.
}
\description{
Description of what getFreq_Alleles does.
}
\examples{
\dontrun{
# Example usage of getFreq_Alleles
result <- getFreq_Alleles(...)
}
}
