% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation_Functions.R
\name{getGenoData_API}
\alias{getGenoData_API}
\title{Title of getGenoData_API}
\usage{
getGenoData_API(FiltGeno)
}
\arguments{
\item{FiltGeno}{Description of FiltGeno.}
}
\value{
Description of what is returned.
}
\description{
Description of what getGenoData_API does.
}
\examples{
\dontrun{
# Example usage of getGenoData_API
result <- getGenoData_API(...)
}
}
