% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getGenoDiff}
\alias{getGenoDiff}
\title{Title of getGenoDiff}
\usage{
getGenoDiff(Geno1, Geno2)
}
\arguments{
\item{Geno1}{Description of Geno1.}

\item{Geno2}{Description of Geno2.}
}
\value{
Description of what is returned.
}
\description{
Description of what getGenoDiff does.
}
\examples{
\dontrun{
# Example usage of getGenoDiff
result <- getGenoDiff(...)
}
}
