% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getGenoImp1Stats}
\alias{getGenoImp1Stats}
\title{Title of getGenoImp1Stats}
\usage{
getGenoImp1Stats(Geno_DF, GenoImp_DF1)
}
\arguments{
\item{Geno_DF}{Description of Geno_DF.}

\item{GenoImp_DF1}{Description of GenoImp_DF1.}
}
\value{
Description of what is returned.
}
\description{
Description of what getGenoImp1Stats does.
}
\examples{
\dontrun{
# Example usage of getGenoImp1Stats
result <- getGenoImp1Stats(...)
}
}
