% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getGenoQCStats}
\alias{getGenoQCStats}
\title{Title of getGenoQCStats}
\usage{
getGenoQCStats(GenoT)
}
\arguments{
\item{GenoT}{Description of GenoT.}
}
\value{
Description of what is returned.
}
\description{
Description of what getGenoQCStats does.
}
\examples{
\dontrun{
# Example usage of getGenoQCStats
result <- getGenoQCStats(...)
}
}
