% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getGenoQCStatsFilt1}
\alias{getGenoQCStatsFilt1}
\title{Title of getGenoQCStatsFilt1}
\usage{
getGenoQCStatsFilt1(GenoT, GenoFilt1T, missSitesTH, MAFTH)
}
\arguments{
\item{GenoT}{Description of GenoT.}

\item{GenoFilt1T}{Description of GenoFilt1T.}

\item{missSitesTH}{Description of missSitesTH.}

\item{MAFTH}{Description of MAFTH.}
}
\value{
Description of what is returned.
}
\description{
Description of what getGenoQCStatsFilt1 does.
}
\examples{
\dontrun{
# Example usage of getGenoQCStatsFilt1
result <- getGenoQCStatsFilt1(...)
}
}
