% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getGenoQCStatsFilt2}
\alias{getGenoQCStatsFilt2}
\title{Title of getGenoQCStatsFilt2}
\usage{
getGenoQCStatsFilt2(GenoFilt1T, GenoFilt2T, missIndTH)
}
\arguments{
\item{GenoFilt1T}{Description of GenoFilt1T.}

\item{GenoFilt2T}{Description of GenoFilt2T.}

\item{missIndTH}{Description of missIndTH.}
}
\value{
Description of what is returned.
}
\description{
Description of what getGenoQCStatsFilt2 does.
}
\examples{
\dontrun{
# Example usage of getGenoQCStatsFilt2
result <- getGenoQCStatsFilt2(...)
}
}
