% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getGenoTas_to_DF}
\alias{getGenoTas_to_DF}
\title{Title of getGenoTas_to_DF}
\usage{
getGenoTas_to_DF(tasGeno)
}
\arguments{
\item{tasGeno}{Description of tasGeno.}
}
\value{
Description of what is returned.
}
\description{
Description of what getGenoTas_to_DF does.
}
\examples{
\dontrun{
# Example usage of getGenoTas_to_DF
result <- getGenoTas_to_DF(...)
}
}
