% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genomic_Kinship_Functions.R
\name{getGmat}
\alias{getGmat}
\title{Compute genomic relationship matrix (G matrix)}
\usage{
getGmat(M_clean, method = "VanRaden")
}
\arguments{
\item{M_clean}{Numeric matrix. QC-filtered marker matrix from getQCFilteredM.}

\item{method}{Character. Estimation method (default "VanRaden").}
}
\value{
A list with elements: G (matrix), msg (character).
}
\description{
Compute genomic relationship matrix (G matrix)
}
