% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation_Functions.R
\name{getImpGenoData_API}
\alias{getImpGenoData_API}
\title{Title of getImpGenoData_API}
\usage{
getImpGenoData_API(vcfIDTab)
}
\arguments{
\item{vcfIDTab}{Description of vcfIDTab.}
}
\value{
Description of what is returned.
}
\description{
Description of what getImpGenoData_API does.
}
\examples{
\dontrun{
# Example usage of getImpGenoData_API
result <- getImpGenoData_API(...)
}
}
