% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation_Functions.R
\name{getImputedData_LDKNNI}
\alias{getImputedData_LDKNNI}
\title{Title of getImputedData_LDKNNI}
\usage{
getImputedData_LDKNNI(FiltGeno, l, k)
}
\arguments{
\item{FiltGeno}{Description of FiltGeno.}

\item{l}{Description of l.}

\item{k}{Description of k.}
}
\value{
Description of what is returned.
}
\description{
Description of what getImputedData_LDKNNI does.
}
\examples{
\dontrun{
# Example usage of getImputedData_LDKNNI
result <- getImputedData_LDKNNI(...)
}
}
