% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation_Functions.R
\name{getImputedData_Num}
\alias{getImputedData_Num}
\title{Title of getImputedData_Num}
\usage{
getImputedData_Num(FiltGeno, nN, Dist)
}
\arguments{
\item{FiltGeno}{Description of FiltGeno.}

\item{nN}{Description of nN.}

\item{Dist}{Description of Dist.}
}
\value{
Description of what is returned.
}
\description{
Description of what getImputedData_Num does.
}
\examples{
\dontrun{
# Example usage of getImputedData_Num
result <- getImputedData_Num(...)
}
}
