% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genomic_Kinship_Functions.R
\name{getKinshipDiag}
\alias{getKinshipDiag}
\title{Kinship diagnostics and duplicate detection}
\usage{
getKinshipDiag(G, duplicate.thr = 0.98)
}
\arguments{
\item{G}{Numeric matrix. Genomic relationship matrix.}

\item{duplicate.thr}{Numeric. Threshold for calling duplicates (default 0.98).}
}
\value{
ASRgenomics kinship.diagnostics list with added msg element.
}
\description{
Kinship diagnostics and duplicate detection
}
