% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genomic_Kinship_Functions.R
\name{getKinshipPCA}
\alias{getKinshipPCA}
\title{PCA on genomic relationship matrix}
\usage{
getKinshipPCA(G, ncp = 20)
}
\arguments{
\item{G}{Numeric matrix. Genomic relationship matrix.}

\item{ncp}{Integer. Number of principal components to retain (default 20).}
}
\value{
FactoMineR PCA object.
}
\description{
PCA on genomic relationship matrix
}
