% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getLibStats}
\alias{getLibStats}
\title{Title of getLibStats}
\usage{
getLibStats(buildLib)
}
\arguments{
\item{buildLib}{Description of buildLib.}
}
\value{
Description of what is returned.
}
\description{
Description of what getLibStats does.
}
\examples{
\dontrun{
# Example usage of getLibStats
result <- getLibStats(...)
}
}
