% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Functions.R
\name{getMEPred}
\alias{getMEPred}
\title{Multi-Environment Genomic Prediction (getME_Pred)}
\usage{
getMEPred(
  DT_1_Filt_List,
  genoDat_List,
  traits,
  KG = NULL,
  KE = NULL,
  KMethod = "Linear",
  FitEnvModels = FALSE,
  fixedME = fixME,
  envVar = varEnv,
  IDColsME,
  LocME = LocationME,
  YrME = YearME
)
}
\arguments{
\item{DT_1_Filt_List}{List of phenotype data frames after merging with genotype data.}

\item{genoDat_List}{List of genotype data frames after merging with phenotype data.}

\item{traits}{Character vector/ Character specifying the traits to analyze.}

\item{KG}{Matrix or NULL. Genotypic relationship matrix. Default = NULL, in which case
the kernel is estimated internally. A user-defined matrix can be supplied.}

\item{KE}{Matrix or NULL. Environmental relationship matrix. Default = NULL.
Required if `FitEnvModels = TRUE`.}

\item{KMethod}{Character. Kernel method for estimating genotypic and
genotype × environment kernels. Default = "Linear". Option: "Gaussian".}

\item{FitEnvModels}{Logical. Whether to include environmental covariates in the model.
Default = FALSE. If TRUE, a user-provided `KE` must be supplied.}

\item{fixedME}{Character. Variable specifying fixed effects in the model.}

\item{envVar}{Character. Variable specifying the environmental factor.}

\item{IDColsME}{Character vector of ID columns (e.g., "Strain", "Location", "Strain-Location").}

\item{LocME}{Character vector. Locations to include. Default = "All" to use all locations,
or specify one or more locations.}

\item{YrME}{Character vector. Years to include. Default = "All" to use all years,
or specify one or more years.}
}
\value{
A list containing results from prediction, including predicted values,
  model fit statistics, variance components and performance metrics..
}
\description{
Performs  genomic prediction in multi-environment trials,
allowing flexible kernel estimation, environmental modeling, and user-defined
cross-validation schemes.
}
\examples{
\dontrun{
# Example usage of getMEPred
result <- getMEPred(...)
}
}
