% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getMergedData}
\alias{getMergedData}
\title{Title of getMergedData}
\usage{
getMergedData(gt2d, Pheno, testIDs)
}
\arguments{
\item{gt2d}{Description of gt2d.}

\item{Pheno}{Description of Pheno.}

\item{testIDs}{Description of testIDs.}
}
\value{
Description of what is returned.
}
\description{
Description of what getMergedData does.
}
\examples{
\dontrun{
# Example usage of getMergedData
result <- getMergedData(...)
}
}
