% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getMergedDataME}
\alias{getMergedDataME}
\title{Title of getMergedDataME}
\usage{
getMergedDataME(phData, genoImp, TargetIDs)
}
\arguments{
\item{phData}{Description of phData.}

\item{genoImp}{Description of genoImp.}

\item{TargetIDs}{Description of TargetIDs.}
}
\value{
Description of what is returned.
}
\description{
Description of what getMergedDataME does.
}
\examples{
\dontrun{
# Example usage of getMergedDataME
result <- getMergedDataME(...)
}
}
