% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getPhenoDistbnPlots}
\alias{getPhenoDistbnPlots}
\title{Title of getPhenoDistbnPlots}
\usage{
getPhenoDistbnPlots(DT_1_Filt_List, TraitME, nTrt, outDirPath)
}
\arguments{
\item{DT_1_Filt_List}{Description of DT_1_Filt_List.}

\item{TraitME}{Description of TraitME.}

\item{nTrt}{Description of nTrt.}

\item{outDirPath}{Description of outDirPath.}
}
\value{
Description of what is returned.
}
\description{
Description of what getPhenoDistbnPlots does.
}
\examples{
\dontrun{
# Example usage of getPhenoDistbnPlots
result <- getPhenoDistbnPlots(...)
}
}
