% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getPhenoMEData}
\alias{getPhenoMEData}
\title{Title of getPhenoMEData}
\usage{
getPhenoMEData(PhenoME, TraitME, nSelTraits, IDColsME, StrainME)
}
\arguments{
\item{PhenoME}{Description of PhenoME.}

\item{TraitME}{Description of TraitME.}

\item{nSelTraits}{Description of nSelTraits.}

\item{IDColsME}{Description of IDColsME.}

\item{StrainME}{Description of StrainME.}
}
\value{
Description of what is returned.
}
\description{
Description of what getPhenoMEData does.
}
\examples{
\dontrun{
# Example usage of getPhenoMEData
result <- getPhenoMEData(...)
}
}
