% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getPredictionData}
\alias{getPredictionData}
\title{Title of getPredictionData}
\usage{
getPredictionData(Data_Table_Num_List, noCandidates)
}
\arguments{
\item{Data_Table_Num_List}{Description of Data_Table_Num_List.}

\item{noCandidates}{Description of noCandidates.}
}
\value{
Description of what is returned.
}
\description{
Description of what getPredictionData does.
}
\examples{
\dontrun{
# Example usage of getPredictionData
result <- getPredictionData(...)
}
}
