% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getProcessedData}
\alias{getProcessedData}
\title{Title of getProcessedData}
\usage{
getProcessedData(Data_Table_Num_List, trait)
}
\arguments{
\item{Data_Table_Num_List}{Description of Data_Table_Num_List.}

\item{trait}{Description of trait.}
}
\value{
Description of what is returned.
}
\description{
Description of what getProcessedData does.
}
\examples{
\dontrun{
# Example usage of getProcessedData
result <- getProcessedData(...)
}
}
