% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getProcessedData_NUST_withFilters}
\alias{getProcessedData_NUST_withFilters}
\title{Title of getProcessedData_NUST_withFilters}
\usage{
getProcessedData_NUST_withFilters(gt2d, NUST_BLUEs, NUST_Meta_Table)
}
\arguments{
\item{gt2d}{Description of gt2d.}

\item{NUST_BLUEs}{Description of NUST_BLUEs.}

\item{NUST_Meta_Table}{Description of NUST_Meta_Table.}
}
\value{
Description of what is returned.
}
\description{
Description of what getProcessedData_NUST_withFilters does.
}
\examples{
\dontrun{
# Example usage of getProcessedData_NUST_withFilters
result <- getProcessedData_NUST_withFilters(...)
}
}
