% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opt_TS_Functions.r
\name{getRandomTS}
\alias{getRandomTS}
\title{Generate a Random Training Subset}
\usage{
getRandomTS(
  Data_Table_Num_Filt_List,
  trait,
  nTraits,
  noCandidates,
  nTrainToSelect
)
}
\arguments{
\item{Data_Table_Num_Filt_List}{A list containing two data frames:
- The first data frame (`TrainData_Table_Num_Filt`) represents the training data.
- The second data frame (`TestData_Table_Num_Filt`) represents the test data.}

\item{trait}{A character vector specifying the trait(s) to analyze. Can be a single trait or multiple traits.}

\item{nTraits}{An integer specifying the number of traits in the dataset.}

\item{noCandidates}{An integer specifying the number of candidates to consider for training.}

\item{nTrainToSelect}{An integer specifying the number of training samples to randomly select.}
}
\value{
A list with the following components:
  - `Train_Random`: A character vector containing the names of the randomly selected training samples.
  - `trainRandomIndices`: An integer vector containing the indices of the randomly selected training samples.
}
\description{
The `getRandomTS` function selects a random subset of training samples from a 
given dataset for use in training genomic models. It supports filtering and 
imputing genotypic data for selected candidates.
}
\details{
- The function processes the input data to extract genotypic and phenotypic information.
- Genotypic data is imputed using the `snpQC` function.
- A random subset of training samples is selected from the candidate set.
- The number of candidates and training samples can be controlled using `noCandidates` and `nTrainToSelect`.
}
\examples{
\dontrun{
# Example input data
TrainData <- data.frame(
  ss1 = c(1, 0, 1),
  ss2 = c(0, 1, 0),
  trait1 = c(5.2, 6.1, 4.8)
)
TestData <- data.frame(
  ss1 = c(1, 1),
  ss2 = c(0, 0),
  trait1 = c(5.9, 4.6)
)

Data_Table_Num_Filt_List <- list(TrainData, TestData)

# Generate a random training subset
result <- getRandomTS(
  Data_Table_Num_Filt_List,
  trait = "trait1",
  nTraits = 1,
  noCandidates = 3,
  nTrainToSelect = 2
)

print(result$Train_Random)
print(result$trainRandomIndices)
}

}
