% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_Functions.R
\name{getRankedPredictedValues}
\alias{getRankedPredictedValues}
\title{Title of getRankedPredictedValues}
\usage{
getRankedPredictedValues(
  Data_Table_Num_Filt_List,
  nTraits,
  trait,
  GPModel,
  fixedX = NULL,
  fixedData = NULL,
  optTS = NULL
)
}
\arguments{
\item{Data_Table_Num_Filt_List}{Description of Data_Table_Num_Filt_List.}

\item{nTraits}{Description of nTraits.}

\item{trait}{Description of trait.}

\item{GPModel}{Description of GPModel.}

\item{fixedX=NULL}{Description of fixedX=NULL.}

\item{fixedData=NULL}{Description of fixedData=NULL.}

\item{optTS=NULL}{Description of optTS=NULL.}
}
\value{
Description of what is returned.
}
\description{
Description of what getRankedPredictedValues does.
}
\examples{
\dontrun{
# Example usage of getRankedPredictedValues
result <- getRankedPredictedValues(...)
}
}
