% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getTasObj}
\alias{getTasObj}
\title{Title of getTasObj}
\usage{
getTasObj(infileVCF)
}
\arguments{
\item{infileVCF}{Description of infileVCF.}
}
\value{
Description of what is returned.
}
\description{
Description of what getTasObj does.
}
\examples{
\dontrun{
# Example usage of getTasObj
result <- getTasObj(...)
}
}
