% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opt_TS_Functions.r
\name{getTSComparisonsMT}
\alias{getTSComparisonsMT}
\title{Title of getTSComparisonsMT}
\usage{
getTSComparisonsMT(
  Data_Table_Num_Filt_List,
  Train_STPGA,
  Train_Random,
  trait,
  nTraits,
  testIds
)
}
\arguments{
\item{Data_Table_Num_Filt_List}{Description of Data_Table_Num_Filt_List.}

\item{Train_STPGA}{Description of Train_STPGA.}

\item{Train_Random}{Description of Train_Random.}

\item{trait}{Description of trait.}

\item{nTraits}{Description of nTraits.}

\item{testIds}{Description of testIds.}
}
\value{
Description of what is returned.
}
\description{
Description of what getTSComparisonsMT does.
}
\examples{
\dontrun{
# Example usage of getTSComparisonsMT
result <- getTSComparisonsMT(...)
}
}
