% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genomic_Kinship_Functions.R
\name{getTunedG}
\alias{getTunedG}
\title{Fine-tune G matrix for positive-definiteness (blend or bend)}
\usage{
getTunedG(G, method = "blend", pblend = 0.05)
}
\arguments{
\item{G}{Numeric matrix. Genomic relationship matrix.}

\item{method}{Character. "blend" or "bend" (default "blend").}

\item{pblend}{Numeric. Blend proportion for identity matrix (default 0.05).}
}
\value{
List with: Gb (tuned matrix), msg, check, check_msg.
}
\description{
Fine-tune G matrix for positive-definiteness (blend or bend)
}
