% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{getU}
\alias{getU}
\title{Calculate the U Statistic}
\usage{
getU(M.Pdt, v)
}
\arguments{
\item{M.Pdt}{A square numeric matrix used to transform the vector `v`.}

\item{v}{A numeric vector to be projected and analyzed.}
}
\value{
A numeric value representing the U statistic, calculated as:
  \deqn{U = \frac{(v^\top M.Pdt^\top M.Pdt v)}{(v^\top v)}}
  where `v.hat = M.Pdt %*% v`.
}
\description{
The `getU` function calculates the U statistic, which is a measure of the 
relative projection of a vector `v` transformed by a matrix `M.Pdt`.
}
\details{
The U statistic is used to assess the proportion of the vector's variance 
that is captured by its projection through the matrix `M.Pdt`.
}
\examples{
\dontrun{
# Example usage
M <- matrix(c(2, 0, 0, 1), nrow = 2) # Example matrix
v <- c(1, 2) # Example vector
getU(M, v)
}

}
