% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{numMAF}
\alias{numMAF}
\title{Title of numMAF}
\usage{
numMAF(Geno, MAFTh)
}
\arguments{
\item{Geno}{Description of Geno.}

\item{MAFTh}{Description of MAFTh.}
}
\value{
Description of what is returned.
}
\description{
Description of what numMAF does.
}
\examples{
\dontrun{
# Example usage of numMAF
result <- numMAF(...)
}
}
