% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{numMissInd}
\alias{numMissInd}
\title{Title of numMissInd}
\usage{
numMissInd(Geno, missIndTH)
}
\arguments{
\item{Geno}{Description of Geno.}

\item{missIndTH}{Description of missIndTH.}
}
\value{
Description of what is returned.
}
\description{
Description of what numMissInd does.
}
\examples{
\dontrun{
# Example usage of numMissInd
result <- numMissInd(...)
}
}
