% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{numMissSites}
\alias{numMissSites}
\title{Title of numMissSites}
\usage{
numMissSites(GenoT, missSitesTH)
}
\arguments{
\item{GenoT}{Description of GenoT.}

\item{missSitesTH}{Description of missSitesTH.}
}
\value{
Description of what is returned.
}
\description{
Description of what numMissSites does.
}
\examples{
\dontrun{
# Example usage of numMissSites
result <- numMissSites(...)
}
}
