% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enviromics_Functions.R
\name{plotEnvRel}
\alias{plotEnvRel}
\title{Title of plotEnvRel}
\usage{
plotEnvRel(KE, outDirPath)
}
\arguments{
\item{KE}{Description of KE.}

\item{outDirPath}{Description of outDirPath.}
}
\value{
Description of what is returned.
}
\description{
Description of what plotEnvRel does.
}
\examples{
\dontrun{
# Example usage of plotEnvRel
result <- plotEnvRel(...)
}
}
