% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enviromics_Functions.R
\name{progress}
\alias{progress}
\title{Title of progress}
\usage{
progress(
  min = 0,
  max = 100,
  leftd = "|",
  rightd = "|",
  char = "=",
  style = 2,
  width = getOption("width"),
  time = Sys.time()
)
}
\arguments{
\item{min}{= 0 Description of min = 0.}

\item{max}{= 100 Description of max = 100.}

\item{leftd}{= "|" Description of leftd = "|".}

\item{rightd}{= "|" Description of rightd = "|".}

\item{char}{= "=" Description of char = "=".}

\item{style}{= 2 Description of style = 2.}

\item{width}{= getOption("width") Description of width = getOption("width").}

\item{time}{= Sys.time() Description of time = Sys.time().}
}
\value{
Description of what is returned.
}
\description{
Description of what progress does.
}
\examples{
\dontrun{
# Example usage of progress
result <- progress(...)
}
}
