% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genomic_Kinship_Functions.R
\name{removeDuplicatesFromG}
\alias{removeDuplicatesFromG}
\title{Remove duplicate individuals from G matrix (greedy)}
\usage{
removeDuplicatesFromG(G, list.duplicate)
}
\arguments{
\item{G}{Numeric matrix. Genomic relationship matrix.}

\item{list.duplicate}{Data frame. Duplicate pairs from getKinshipDiag.}
}
\value{
List with: G_clean, removed, n_pairs, msg.
}
\description{
Remove duplicate individuals from G matrix (greedy)
}
