% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enviromics_Functions.R
\name{run_progress}
\alias{run_progress}
\title{Title of run_progress}
\usage{
run_progress(pb, actual, text = "", digits = 0, sleep = 0)
}
\arguments{
\item{pb}{Description of pb.}

\item{actual}{Description of actual.}

\item{text}{= "" Description of text = "".}

\item{digits}{= 0 Description of digits = 0.}

\item{sleep}{= 0 Description of sleep = 0.}
}
\value{
Description of what is returned.
}
\description{
Description of what run_progress does.
}
\examples{
\dontrun{
# Example usage of run_progress
result <- run_progress(...)
}
}
