% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enviromics_Functions.R
\name{sec_to_hms}
\alias{sec_to_hms}
\title{Title of sec_to_hms}
\usage{
sec_to_hms(seconds)
}
\arguments{
\item{seconds}{Description of seconds.}
}
\value{
Description of what is returned.
}
\description{
Description of what sec_to_hms does.
}
\examples{
\dontrun{
# Example usage of sec_to_hms
result <- sec_to_hms(...)
}
}
