% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enviromics_Functions.R
\name{split_chunk}
\alias{split_chunk}
\title{Title of split_chunk}
\usage{
split_chunk(vec, length)
}
\arguments{
\item{vec}{Description of vec.}

\item{length}{Description of length.}
}
\value{
Description of what is returned.
}
\description{
Description of what split_chunk does.
}
\examples{
\dontrun{
# Example usage of split_chunk
result <- split_chunk(...)
}
}
