% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Functions.R
\name{stdizePhVarNames}
\alias{stdizePhVarNames}
\title{Standardize ME phenotype column names for downstream pipeline use}
\usage{
stdizePhVarNames(
  PhenoME,
  IDColsME,
  StrainME,
  EnvVarME,
  LocMEColVar,
  YrMEColVar
)
}
\arguments{
\item{PhenoME}{Data frame. Raw ME phenotype table.}

\item{IDColsME}{Character vector. All ID columns in PhenoME.}

\item{StrainME}{Character. Column name for genotype/strain ID.}

\item{EnvVarME}{Character. Column name for the environment factor.}

\item{LocMEColVar}{Character or NULL. Location column name.}

\item{YrMEColVar}{Character or NULL. Year column name.}
}
\value{
List: [[1]] modified phenotype data frame, [[2]] updated IDColsME, [[3]] standardized env variable name.
}
\description{
Standardize ME phenotype column names for downstream pipeline use
}
