% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Functions.R
\name{summarize_ME_CV_fits}
\alias{summarize_ME_CV_fits}
\title{summarize_ME_CV_fits}
\usage{
summarize_ME_CV_fits(fit_obj, cvMet = CVMet)
}
\arguments{
\item{fit_obj}{output from fitMEModels_CV()}

\item{DT}{data table used for fitting the models in fitMEModels_CV}

\item{IDColsList}{a vector containing the ID cols in ME data table}
}
\value{
A list containing various metrics: predictive ability, RMSE,
  variance components, fit values, and summary of these metrics.
}
\description{
Summarize output from fitMEModels_CV_fits
}
\examples{
\dontrun{
result <- summarize_ME_CV_fits(fit_obj, cvMet = "CV1")
}
}
