% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enviromics_Functions.R
\name{syncEnvPhenoDat}
\alias{syncEnvPhenoDat}
\title{Title of syncEnvPhenoDat}
\usage{
syncEnvPhenoDat(KE, LocCoord, OtherLoc)
}
\arguments{
\item{KE}{Description of KE.}

\item{LocCoord}{Description of LocCoord.}

\item{OtherLoc}{Description of OtherLoc.}
}
\value{
Description of what is returned.
}
\description{
Description of what syncEnvPhenoDat does.
}
\examples{
\dontrun{
# Example usage of syncEnvPhenoDat
result <- syncEnvPhenoDat(...)
}
}
