% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{VCFtoDF}
\alias{VCFtoDF}
\title{Title of VCFtoDF}
\usage{
VCFtoDF(infile)
}
\arguments{
\item{infile}{Description of infile.}
}
\value{
Description of what is returned.
}
\description{
Description of what VCFtoDF does.
}
\examples{
\dontrun{
# Example usage of VCFtoDF
result <- VCFtoDF(...) 
}
}
