% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Functions.R
\name{Vcomp.BGGE}
\alias{Vcomp.BGGE}
\title{Title of Vcomp.BGGE}
\usage{
Vcomp.BGGE(model, env, gid, digits = digits, alfa = 0.1)
}
\arguments{
\item{model}{Description of model.}

\item{env}{Description of env.}

\item{gid}{Description of gid.}

\item{digits}{= digits Description of digits = digits.}

\item{alfa}{= 0.1 Description of alfa = 0.1.}
}
\value{
Description of what is returned.
}
\description{
Description of what Vcomp.BGGE does.
}
\examples{
\dontrun{
# Example usage of Vcomp.BGGE
result <- Vcomp.BGGE(...)
}
}
