% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{writeCVOutTable}
\alias{writeCVOutTable}
\title{Title of writeCVOutTable}
\usage{
writeCVOutTable(CVROut, type, outDirPath)
}
\arguments{
\item{CVROut}{Description of CVROut.}

\item{type}{Description of type.}

\item{outDirPath}{Description of outDirPath.}
}
\value{
Description of what is returned.
}
\description{
Description of what writeCVOutTable does.
}
\examples{
\dontrun{
# Example usage of writeCVOutTable
result <- writeCVOutTable(...)
}
}
