% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS_Pipeline_FnsApp.r
\name{writeGPOutTable}
\alias{writeGPOutTable}
\title{Title of writeGPOutTable}
\usage{
writeGPOutTable(GPOut, type, outDirPath)
}
\arguments{
\item{GPOut}{Description of GPOut.}

\item{type}{Description of type.}

\item{outDirPath}{Description of outDirPath.}
}
\value{
Description of what is returned.
}
\description{
Description of what writeGPOutTable does.
}
\examples{
\dontrun{
# Example usage of writeGPOutTable
result <- writeGPOutTable(...)
}
}
