% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Functions.R
\name{getME_CV}
\alias{getME_CV}
\title{Title of getME_CV}
\usage{
getME_CV(
  DT_1_Filt_List,
  genoDat_List,
  traits,
  KG = NULL,
  KE = NULL,
  CVMet,
  factVar,
  K,
  NIter,
  KMethod = "Linear",
  FitEnvModels = FALSE,
  fixedME = fixME,
  envVar = varEnv,
  IDColsME,
  IDColME,
  LocME = LocationME,
  YrME = YearME
)
}
\arguments{
\item{DT_1_Filt_List}{Description of DT_1_Filt_List.}

\item{genoDat_List}{Description of genoDat_List.}

\item{traits}{Description of traits.}

\item{CVMet}{Description of CVMet.}

\item{factVar}{Description of factVar.}

\item{K}{Description of K.}

\item{NIter}{Description of NIter.}

\item{IDColsME}{Description of IDColsME.}

\item{IDColME}{Description of IDColME.}

\item{KG=NULL}{Description of KG=NULL.}

\item{KE=NULL}{Description of KE=NULL.}

\item{KMethod="Linear"}{Description of KMethod="Linear".}

\item{FitEnvModels=FALSE}{Description of FitEnvModels=FALSE.}

\item{fixedME=fixME}{Description of fixedME=fixME.}

\item{envVar=varEnv}{Description of envVar=varEnv.}

\item{LocME=LocationME}{Description of LocME=LocationME.}

\item{YrME=YearME}{Description of YrME=YearME.}
}
\value{
Description of what is returned.
}
\description{
Description of what getME_CV does.
}
\examples{
\dontrun{
# Example usage of getME_CV
result <- getME_CV(...)
}
}
