% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Functions.R
\name{getMEPred}
\alias{getMEPred}
\title{Title of getMEPred}
\usage{
getMEPred(
  DT_1_Filt_List,
  genoDat_List,
  traits,
  KG = NULL,
  KE = NULL,
  KMethod = "Linear",
  FitEnvModels = FALSE,
  fixedME = fixME,
  envVar = varEnv,
  IDColsME,
  LocME = LocationME,
  YrME = YearME
)
}
\arguments{
\item{DT_1_Filt_List}{Description of DT_1_Filt_List.}

\item{genoDat_List}{Description of genoDat_List.}

\item{traits}{Description of traits.}

\item{IDColsME}{Description of IDColsME.}

\item{KG=NULL}{Description of KG=NULL.}

\item{KE=NULL}{Description of KE=NULL.}

\item{KMethod="Linear"}{Description of KMethod="Linear".}

\item{FitEnvModels=FALSE}{Description of FitEnvModels=FALSE.}

\item{fixedME=fixME}{Description of fixedME=fixME.}

\item{envVar=varEnv}{Description of envVar=varEnv.}

\item{LocME=LocationME}{Description of LocME=LocationME.}

\item{YrME=YearME}{Description of YrME=YearME.}
}
\value{
Description of what is returned.
}
\description{
Description of what getMEPred does.
}
\examples{
\dontrun{
# Example usage of getMEPred
result <- getMEPred(...)
}
}
