% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MT_Functions.R
\name{getMTCVR}
\alias{getMTCVR}
\title{Title of getMTCVR}
\usage{
getMTCVR(Data_Table_Num_Filt_List, trait, nTraits, k, nIter)
}
\arguments{
\item{Data_Table_Num_Filt_List}{Description of Data_Table_Num_Filt_List.}

\item{trait}{Description of trait.}

\item{nTraits}{Description of nTraits.}

\item{k}{Description of k.}

\item{nIter}{Description of nIter.}
}
\value{
Description of what is returned.
}
\description{
Description of what getMTCVR does.
}
\examples{
\dontrun{
# Example usage of getMTCVR
result <- getMTCVR(...)
}
}
