% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MT_Functions.R
\name{getRankedPredictedValuesMT}
\alias{getRankedPredictedValuesMT}
\title{Title of getRankedPredictedValuesMT}
\usage{
getRankedPredictedValuesMT(
  Data_Table_Num_Filt_List,
  nTraits,
  trait,
  GPModelMT,
  optTS = NULL
)
}
\arguments{
\item{Data_Table_Num_Filt_List}{Description of Data_Table_Num_Filt_List.}

\item{nTraits}{Description of nTraits.}

\item{trait}{Description of trait.}

\item{GPModelMT}{Description of GPModelMT.}

\item{optTS=NULL}{Description of optTS=NULL.}
}
\value{
Description of what is returned.
}
\description{
Description of what getRankedPredictedValuesMT does.
}
\examples{
\dontrun{
# Example usage of getRankedPredictedValuesMT
result <- getRankedPredictedValuesMT(...)
}
}
