% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME_Functions.R
\name{getTstIndices_CV_Data}
\alias{getTstIndices_CV_Data}
\title{Title of getTstIndices_CV_Data}
\usage{
getTstIndices_CV_Data(DT_2B, CVMet, nIter, k)
}
\arguments{
\item{DT_2B}{Description of DT_2B.}

\item{CVMet}{Description of CVMet.}

\item{nIter}{Description of nIter.}

\item{k}{Description of k.}
}
\value{
Description of what is returned.
}
\description{
Description of what getTstIndices_CV_Data does.
}
\examples{
\dontrun{
# Example usage of getTstIndices_CV_Data
result <- getTstIndices_CV_Data(...)
}
}
